# [PERF] stock, mrp: speedup _compute_quantities