--- a/addons/mrp/models/product.py
+++ b/addons/mrp/models/product.py
@@ -276,8 +276,12 @@ class ProductProduct(models.Model):
         # pre-compute bom lines and identify missing kit components to prefetch
         bom_sub_lines_per_kit = {}
         prefetch_component_ids = set()
-        for product in bom_kits:
-            __, bom_sub_lines = bom_kits[product].explode(product, 1)
+        # memoize explode calls as we are using bom_sub_lines that depends only on the bom.
+        exploded_values = {}
+        for product, bom in bom_kits.items():
+            if not exploded_values.get(bom):
+                exploded_values[bom] = bom_kits[product].explode(product, 1)
+            __, bom_sub_lines = exploded_values[bom]
             bom_sub_lines_per_kit[product] = bom_sub_lines
             for bom_line, __ in bom_sub_lines:
                 if bom_line.product_id.id not in qties: